requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.10
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
google-generativeai==0.8.3
pypdf==5.0.1
Pillow==10.4.0
extruct==0.18.0
w3lib==2.2.1
